+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Many return impl Trait changed lifetimes slightly, some replaced with doc hidden types. In practice all code should still just compile,
   but these are breaking changes.
 * Some public `hashbrown::{HashMap, HashSet}` API replaced with the `std::collections` equivalent.
+* Remove deprecated `zng::text::justify` that was replaced with `justify_mode` in 0.13.11.
 
 # 0.13.12